refactor,chore: refactoring app structure - init 1st stage
