+++ src/components/ScreenWrapper.tsx
@@ -1,24 +1,16 @@
-import { View, Text, ImageBackground } from "react-native";
+import { View } from "react-native";
+import Tabs from "./Tabs";
+import Header from "./Header";
 
-interface ScreenWrapperProps {
-  children: React.ReactNode;
-}
-
-const BackgroundImage = require("../../assets/images/bg-mobile-dark.jpg");
-
-const ScreenWrapper = ({ children }: ScreenWrapperProps) => {
+const ScreensWrapper = () => {
   return (
-    <View className="h-full">
-      <ImageBackground source={BackgroundImage} style={{ flex: 0.3 }}>
-        <Text className="text-white mt-[60] tracking-[18] font-bold text-3xl uppercase text-center">
-          Todo
-        </Text>
-      </ImageBackground>
-      <View className="h-full p-2" style={{ flex: 0.7 }}>
-        {children}
+    <View>
+      <Header />
+      <View>
+        <Tabs />
       </View>
     </View>
   );
 };
 
-export default ScreenWrapper;
+export default ScreensWrapper;
